+++ src/bots/slack/controller.ts
@@ -1,15 +1,17 @@
-import { SlackAdapter, SlackBotWorker } from 'botbuilder-adapter-slack';
-import { Botkit, BotkitMessage } from 'botkit';
-import { Response } from 'express';
+import { SlackAdapter } from 'botbuilder-adapter-slack';
+import { Botkit } from 'botkit';
+
 import { storage } from './storage';
 import { SLACK_WEBHOOK } from '../../constants';
 import { app } from '../../server';
 // Conversations
 import { debugConversation } from './modules/debug/debug.conversation';
 import { introductionDialog } from './modules/introduction-example.ts/introduction';
+import { slashCommandExample } from './modules/slash-command-example.ts/slash-command';
 
 export const initSlackController = (adapter: SlackAdapter) => {
   const controller = new Botkit({
+    dialogStateProperty: `botbuilder-adapter-slack`,
     adapter,
     webhook_uri: SLACK_WEBHOOK,
     storage,
@@ -23,6 +25,9 @@ export const initSlackController = (adapter: SlackAdapter) => {
 
     // Example that asks a user their name and favorite color
     introductionDialog(controller);
+
+    // Example for handling a Slash Command
+    slashCommandExample(controller);
     console.log('*****\nSlackbot Modules Loaded!\n*****');
   });
 };